+++ R/textstat.R
@@ -19,31 +19,6 @@
 #'   dfm_trim dfm_match featnames as.dfm dfm_remove
 #' @export
 #' @seealso [tokens_select()] and [textstat_keyness()]
-#' @examples
-#' \donttest{
-#' require(quanteda)
-#' con <- url("https://bit.ly/2GZwLcN", "rb")
-#' corp <- readRDS(con)
-#' close(con)
-#' corp <- corpus_reshape(corp, 'sentences')
-#' toks <- tokens(corp, remove_punct = TRUE)
-#' toks <- tokens_remove(toks, stopwords("en"))
-#'
-#' # economy keywords
-#' eco <- char_context(toks, 'econom*')
-#' head(eco, 20)
-#'
-#' tstat_eco <- textstat_context(toks, 'econom*')
-#' head(tstat_eco)
-#'
-#' # politics keywords
-#' pol <- char_context(toks, 'politi*')
-#' head(pol, 20)
-#'
-#' # politics keywords
-#' tstat_pol <- textstat_context(toks, 'politi*')
-#' head(tstat_pol)
-#' }
 textstat_context <- function(x, pattern, valuetype = c("glob", "regex", "fixed"),
                              case_insensitive = TRUE, window = 10, min_count = 10,
                              remove_pattern = TRUE, n = 1, skip = 0, ...) {
@@ -97,17 +72,4 @@ char_context <- function(x, pattern, valuetype = c("glob", "regex", "fixed"),
                                n = n, skip = skip)
     result <- result[result[[2]] > 0 & result$p < p,]
     return(result$feature)
-}
-
-#' @noRd
-#' @keywords internal
-#' @export
-char_keyness <- function(x, pattern, valuetype = c("glob", "regex", "fixed"),
-                         case_insensitive = TRUE, window = 10, min_count = 10,
-                         remove_pattern = TRUE, p = 0.001, n = 1, skip = 0) {
-    .Deprecated(msg = "char_keyness() is deprecated. Use char_context() instead.")
-    char_context(x, pattern = pattern, valuetype = valuetype,
-                 case_insensitive = case_insensitive, window = window,
-                 min_count = min_count, remove_pattern = remove_pattern, p = p,
-                 n = n, skip = skip)
 }