+++ tests/testthat/test-textmodel.R
@@ -112,18 +112,26 @@ test_that("predict.textmodel_lss is working", {
     expect_equal(length(pred3$density), ndoc(toks_test))
     expect_null(names(pred3$density))
 
-    pred4 <- predict(lss_test, rescaling = FALSE)
+    pred4 <- predict(lss_test, rescale = FALSE)
     expect_identical(names(pred4), docnames(toks_test))
     expect_equal(as.numeric(scale(pred4)), unname(pred1))
 
     pred5 <- predict(lss_test, se_fit = TRUE, density = TRUE)
     expect_equal(names(pred5), c("fit", "se.fit", "n", "density"))
 
-    pred6 <- predict(lss_test, rescaling = FALSE, min_n = 2)
+    pred6 <- predict(lss_test, rescale = FALSE, min_n = 2)
     expect_true(all(is.na(pred4) == is.na(pred6)))
     expect_true(all(abs(pred6[pred5$n == 1]) < abs(pred4[pred5$n == 1]), na.rm = TRUE))
     expect_true(all(abs(pred6[pred5$n >= 2]) == abs(pred4[pred5$n >= 2]), na.rm = TRUE))
 
+    expect_error(
+        predict(lss_test, rescale = FALSE, min_n = -1),
+        "The value of min_n must be between 0 and Inf"
+    )
+    expect_error(
+        predict(lss_test, rescale = FALSE, min_n = c(0, 1)),
+        "The length of min_n must be 1"
+    )
 })
 
 test_that("density is correct", {
@@ -153,7 +161,7 @@ test_that("calculation of fit and se_fit are correct", {
     lss <- as.textmodel_lss(c("a" = 0.1, "b" = 0.1, "c" = 0.3))
     toks <- tokens(c("a a a", "a b", "a a b c c d e"))
     dfmt <- dfm(toks)
-    pred <- predict(lss, newdata = dfmt, se_fit = TRUE, rescaling = FALSE)
+    pred <- predict(lss, newdata = dfmt, se_fit = TRUE, rescale = FALSE)
 
     expect_equal(pred$fit[1], c(text1 = 0.10))
     expect_equal(pred$fit[2], c(text2 = 0.10))
@@ -231,11 +239,23 @@ test_that("simil_method works", {
 
 
 test_that("include_data is working", {
-    dfmt <- dfm_group(dfm(toks_test))
-    lss <- textmodel_lss(dfmt, seedwords("pos-neg"), include_data = TRUE, k = 10)
-    lss_nd <- textmodel_lss(dfmt, seedwords("pos-neg"), include_data = FALSE, k = 10)
+
+    skip_on_cran() # takes to much time
+
+    dfmt <- dfm(toks_test)
+    lss <- textmodel_lss(dfmt, seedwords("pos-neg"), k = 10, include_data = TRUE)
+    lss_nd <- textmodel_lss(dfmt, seedwords("pos-neg"), k = 10, include_data = FALSE)
     expect_error(predict(lss_nd), "The model includes no data")
     expect_identical(predict(lss), predict(lss_nd, newdata = dfmt))
+
+    lss_gd <- textmodel_lss(dfmt, seedwords("pos-neg"), k = 10,
+                            include_data = TRUE, group_data = TRUE)
+    expect_equal(names(predict(lss_gd)), docnames(dfm_group(dfmt)))
+    expect_warning(
+        textmodel_lss(dfmt, seedwords("pos-neg"), k = 10,
+                      include_data = FALSE, group_data = TRUE),
+        "group_data is ignored when include_data = FALSE"
+    )
 })
 
 test_that("predict.textmodel_lss computes scores correctly", {
@@ -295,6 +315,9 @@ test_that("RSpectra and irlba work", {
 })
 
 test_that("text2vec works", {
+
+    skip_on_cran() # takes to much time
+
     fcmt <- fcm(toks_test)
     lss <- textmodel_lss(fcmt, seedwords("pos-neg"), engine = "rsparse")
     expect_equal(
@@ -333,6 +356,8 @@ test_that("slice argument is working", {
 })
 
 test_that("test smooth_lss", {
+
+    skip_on_cran() # takes to much time
 
     set.seed(1234)
     dfmt <- dfm_sample(dfmt_test, size = 1000)
@@ -400,7 +425,9 @@ test_that("weight_seeds() works", {
 })
 
 test_that("old argument still works", {
-    skip_on_cran()
+
+    skip_on_cran() # takes to much time
+
     suppressWarnings({
         lss <- textmodel_lss(dfmt_test, seed, features = feat_test, k = 300)
     })
@@ -417,7 +444,65 @@ test_that("se_fit is working", {
     lss <- as.textmodel_lss(beta)
     dfmt1 <- dfm(tokens(c("a a a b b", "")))
     dfmt2 <- dfm(tokens(c("a a a b b z z z z z", "")))
-    pred1 <- predict(lss, newdata = dfmt1, rescaling = FALSE, min_n = 10, se_fit = TRUE)
-    pred2 <- predict(lss, newdata = dfmt2, rescaling = FALSE, se_fit = TRUE)
+    pred1 <- predict(lss, newdata = dfmt1, rescale = FALSE, min_n = 10, se_fit = TRUE)
+    pred2 <- predict(lss, newdata = dfmt2, rescale = FALSE, se_fit = TRUE)
     expect_identical(pred1, pred2)
+})
+
+test_that("cut is working", {
+
+    skip_on_cran() # takes to much time
+
+    p0 <- predict(lss_test, rescale = TRUE, min_n = 10)
+    p1 <- predict(lss_test, cut = 0.5, rescale = TRUE)
+    expect_true(min(p1, na.rm = TRUE) < -1)
+    expect_true(max(p1, na.rm = TRUE) > 1)
+    expect_equal(cor(p0, p1, use = "pair"), 0.59, tolerance = 0.01)
+
+    p2 <- predict(lss_test, cut = 0.5, rescale = FALSE)
+    expect_true(min(p2, na.rm = TRUE) >= -1)
+    expect_true(max(p2, na.rm = TRUE) <= 1)
+    expect_equal(cor(p0, p2, use = "pair"), 0.59, tolerance = 0.01)
+
+    p3 <- predict(lss_test, cut = 0.5, rescale = FALSE, min_n = 10)
+    expect_true(min(p3, na.rm = TRUE) >= -1)
+    expect_true(max(p3, na.rm = TRUE) <= 1)
+    expect_equal(cor(p0, p3, use = "pair"), 0.73, tolerance = 0.01)
+
+    p4 <- predict(lss_test, cut = 0.75, rescale = FALSE, min_n = 10)
+    expect_true(min(p4, na.rm = TRUE) >= -1)
+    expect_true(max(p4, na.rm = TRUE) <= 1)
+    expect_equal(cor(p0, p4, use = "pair"), 0.33, tolerance = 0.01)
+
+    p5 <- predict(lss_test, cut = c(0.25, 0.75), rescale = FALSE, min_n = 10)
+    expect_true(min(p5, na.rm = TRUE) >= -1)
+    expect_true(max(p5, na.rm = TRUE) <= 1)
+    expect_equal(cor(p0, p5, use = "pair"), 0.77, tolerance = 0.01)
+
+    p6 <- predict(lss_test, cut = c(0.75, 0.25), rescale = FALSE, min_n = 10)
+    expect_identical(p5, p6)
+
+    expect_error(
+        predict(lss_test, cut = 1.5),
+        "The value of cut must be between 0 and 1"
+    )
+    expect_error(
+        predict(lss_test, cut = -0.1),
+        "The value of cut must be between 0 and 1"
+    )
+    expect_error(
+        predict(lss_test, cut = c(0.1, 0.5, 0.9)),
+        "The length of cut must be between 1 and 2"
+    )
+})
+
+test_that("rescaling still works", {
+
+    expect_warning({
+        p1 <- predict(lss_test, rescaling = TRUE)
+    })
+    expect_silent({
+        p2 <- predict(lss_test, rescale = TRUE)
+    })
+    expect_identical(p1, p2)
 })